+++ NEWS.md
@@ -8,6 +8,7 @@ New:
 
 Bug:
 
+* `avg_predictions()` did not respect the `wts` argument when using `variables` to create counterfactual predictions. Thanks to @strengejacke for report #1596.
 * Binary variables treated as integers by `datagrid()` with `newdata=` instead of `model=`. Issue #1584.
 * Better labelling in some `brms` models with categorical outcomes. Thanks to @mattansb for a proposed solution and to @strengejacke for report #1392.
 * Improve printing of grouped hypotheses. Issue #1579.